Crate rome_json_syntax

Modules

kind

Generated file, do not edit by hand, see xtask/codegen

macros

Generated file, do not edit by hand, see xtask/codegen

Macros

T

Utility macro for creating a SyntaxKind through simple macro syntax

map_syntax_node

Reconstruct an AstNode from a SyntaxNode

Structs

DebugSyntaxElementChildren

JsonArray

JsonArrayElementList

JsonArrayFields

JsonBoolean

JsonBooleanFields

JsonLanguage

JsonMember

JsonMemberFields

JsonMemberList

JsonNull

JsonNullFields

JsonNumber

JsonNumberFields

JsonObject

JsonObjectFields

JsonRoot

JsonRootFields

JsonString

JsonStringFields

JsonUnknown

TextRange

A range in text, represented as a pair of TextSize.

TextSize

A measure of text length. Also, equivalently, an index into text.

Enums

JsonSyntaxKind

The kind of syntax node, e.g. IDENT, FUNCTION_KW, or FOR_STMT.

JsonValue

TokenAtOffset

There might be zero, one or two leaves at a given offset.

TriviaPieceKind

WalkEvent

WalkEvent describes tree walking process.

Traits

TextLen

Primitives with a textual length that can be passed to TextSize::of.

Type Definitions

JsonSyntaxElement

JsonSyntaxElementChildren

JsonSyntaxList

JsonSyntaxNode

JsonSyntaxNodeChildren

JsonSyntaxToken